# Proposal: add UnpackBlock to BlockUnpacker interface