This PR is not intended to be merged, it is just a proposal with code examples.

Although there is a lot of code, the proposal is only about expanding the BlockUnpacker interface and implementing the new method in `pack/{ee,eeintegrity,plain}`.

# Proposal

Add the following method to the `upspin.BlockUnpacker` interface:
```go
        // UnpackBlock takes ciphertext and the block number and returns the
	// cleartext. If appropriate, the result is verified as correct according
	// to the block's Packdata.
	//
	// UnpackBlock, different from Upack, is not affected by NextBlock.
	//
	// The cleartext slice remains valid until the next call to Unpack or
	// UnpackBlock.
	UnpackBlock(ciphertext []byte, n int) (cleartext []byte, err error)
```

UnpackBlock works like Unpack but is independent from method NextBlock.
The fact that those methods are independent make possible (or at least more ergonomic) to call those functions concurrently.

# Motivation

Why do I need to call NextBlock and UnpackBlock concurrently?

Performance.
I have a remote Upspin server that uses a S3 like storage, so there is the latency from my laptop to the Upspin server + the latency from the storage server to get data from the S3 storage. I guess I have a latency of ~100ms to get a block from my Upspin server.

The current `client/clientutil.ReadAll()` function fetch and read each block of 1MB (or other BlockSize) sequentially. So if I read a file of 50MB size I take a penalty of 50 * 100ms = 5s just waiting for latency reasons.

If I could use NextBlock and UnpackBlock concurrently I could fetch N blocks concurrently and then just unpack sequentially, using some sliding window algorithm for example.

`client/clientutil.ReadAll()` is not the only function that could be parallelized I suspect that the client File.Read() could benefit from that too, but I didn't dive deeper.

# Benchmark

As I said in the beginning of this proposal, this PR contains a lot of code that is not directly related to the proposal.

I wrote a proof of concept of a concurrently `ReadAll()` function and a simple Benchmark, both of which are not well written and probably contains bugs, but I think they are useful to show the intuition and the performance gains.

See `client/clientutil/readall.go` and `client/clientutil/benchmark_test.go`.

The benchmark results I got simulating a server with 50ms of latency:
```
goos: darwin
goarch: arm64
pkg: upspin.io/client/clientutil
cpu: Apple M3 Pro
           │   old.txt    │               new.txt               │
           │    sec/op    │   sec/op     vs base                │
ReadAll-11   2694.4m ± 0%   225.0m ± 1%  -91.65% (p=0.000 n=10)

           │   old.txt    │               new.txt                │
           │     B/op     │     B/op      vs base                │
ReadAll-11   316.8Mi ± 0%   276.7Mi ± 0%  -12.65% (p=0.000 n=10)

           │  old.txt   │              new.txt               │
           │ allocs/op  │ allocs/op   vs base                │
ReadAll-11   492.0 ± 1%   351.0 ± 1%  -28.66% (p=0.000 n=10)
```