@n2vi 2025-01-01
The idea seems ok to me; there is no intrinsic reason why block access
can't be parallelized.

On Mon, Dec 23, 2024 at 9:19 PM Rodrigo Schio ***@***.***>
wrote:

> This PR is not intended to be merged, it is just a proposal with code
> examples.
>
> Although there is a lot of code, the proposal is only about expanding the
> BlockUnpacker interface and implementing the new method in
> pack/{ee,eeintegrity,plain}.
> Proposal
>
> Add the following method to the upspin.BlockUnpacker interface:
>
>         // UnpackBlock takes ciphertext and the block number and returns the
> 	// cleartext. If appropriate, the result is verified as correct according
> 	// to the block's Packdata.
> 	//
> 	// UnpackBlock, different from Upack, is not affected by NextBlock.
> 	//
> 	// The cleartext slice remains valid until the next call to Unpack or
> 	// UnpackBlock.
> 	UnpackBlock(ciphertext []byte, n int) (cleartext []byte, err error)
>
> UnpackBlock works like Unpack but is independent from method NextBlock.
> The fact that those methods are independent make possible (or at least
> more ergonomic) to call those functions concurrently.
> Motivation
>
> Why do I need to call NextBlock and UnpackBlock concurrently?
>
> Performance.
> I have a remote Upspin server that uses a S3 like storage, so there is the
> latency from my laptop to the Upspin server + the latency from the storage
> server to get data from the S3 storage. I guess I have a latency of ~100ms
> to get a block from my Upspin server.
>
> The current client/clientutil.ReadAll() function fetch and read each
> block of 1MB (or other BlockSize) sequentially. So if I read a file of 50MB
> size I take a penalty of 50 * 100ms = 5s just waiting for latency reasons.
>
> If I could use NextBlock and UnpackBlock concurrently I could fetch N
> blocks concurrently and then just unpack sequentially, using some sliding
> window algorithm for example.
>
> client/clientutil.ReadAll() is not the only function that could be
> parallelized I suspect that the client File.Read() could benefit from that
> too, but I didn't dive deeper.
> Benchmark
>
> As I said in the beginning of this proposal, this PR contains a lot of
> code that is not directly related to the proposal.
>
> I wrote a proof of concept of a concurrently ReadAll() function and a
> simple Benchmark, both of which are not well written and probably contains
> bugs, but I think they are useful to show the intuition and the performance
> gains.
>
> See client/clientutil/readall.go and client/clientutil/benchmark_test.go.
>
> The benchmark results I got simulating a server with 50ms of latency:
>
> goos: darwin
> goarch: arm64
> pkg: upspin.io/client/clientutil
> cpu: Apple M3 Pro
>            │   old.txt    │               new.txt               │
>            │    sec/op    │   sec/op     vs base                │
> ReadAll-11   2694.4m ± 0%   225.0m ± 1%  -91.65% (p=0.000 n=10)
>
>            │   old.txt    │               new.txt                │
>            │     B/op     │     B/op      vs base                │
> ReadAll-11   316.8Mi ± 0%   276.7Mi ± 0%  -12.65% (p=0.000 n=10)
>
>            │  old.txt   │              new.txt               │
>            │ allocs/op  │ allocs/op   vs base                │
> ReadAll-11   492.0 ± 1%   351.0 ± 1%  -28.66% (p=0.000 n=10)
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/upspin/upspin/pull/671
> Commit Summary
>
>    - 492fd06
>    <https://github.com/upspin/upspin/pull/671/commits/492fd06a5d3323263325b91803e850d2f94985f8>
>    client: fix client/file and client/clientutil tests
>    - 576e818
>    <https://github.com/upspin/upspin/pull/671/commits/576e81808dd48a8c32b4e24c3de2691ad9bb9144>
>    client/clientutil: add BenchmarkReadAll
>    - c7206e3
>    <https://github.com/upspin/upspin/pull/671/commits/c7206e3ec0a77f4991fe05d551305205faf72e47>
>    upspin: add the method UnpackBlock to the BlockUnpacker interface
>    - fba6f49
>    <https://github.com/upspin/upspin/pull/671/commits/fba6f4958394e87318cd060f229a1bc7f7fe0236>
>    client/clientutil: ReadAll request blocks concurrently
>
> File Changes
>
> (8 files <https://github.com/upspin/upspin/pull/671/files>)
>
>    - *M* client/clientutil/all_test.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-c27d471fa18637bc857396a2fd4d6c23f9be161cfa1cdf7bfd99a926a19464ef>
>    (5)
>    - *A* client/clientutil/benchmark_test.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-ec55024fcbdbedac0889efb721156324df23d7c9b7b57b4d3c4ef3fce3a56ae5>
>    (137)
>    - *M* client/clientutil/readall.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-77ad1d41571b88a36e80b440ff1037a7b698f7a8c9cdd7b1ec528b37d5acdb98>
>    (133)
>    - *M* client/file/file_test.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-49515c388bb7e1daa55e79e165b5bef52be3ad196b9f2987cab18729791140bf>
>    (3)
>    - *M* pack/ee/ee.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-04d8286e52b7095c43e1e29ed97ae2f42ac10e3e8776aa219e804fa053732d51>
>    (9)
>    - *M* pack/eeintegrity/eeintegrity.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-087197cfc45bf42db8b795983df4ad055c4d13b86e5f7b9163f8a4eeaa32f943>
>    (7)
>    - *M* pack/plain/plain.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-8630c7a40c6a3f46e6418c482a6a0b59597d3f1ed796a281aacf84a948e73a26>
>    (4)
>    - *M* upspin/upspin.go
>    <https://github.com/upspin/upspin/pull/671/files#diff-2c6b50875dbcbda6192f3fbde6a0d29691aa990e4446cae591e9f2c7e5cf2e33>
>    (15)
>
> Patch Links:
>
>    - https://github.com/upspin/upspin/pull/671.patch
>    - https://github.com/upspin/upspin/pull/671.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/upspin/upspin/pull/671>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACADPOSS6TN4W6KWZVNR2RL2HDVH5AVCNFSM6AAAAABUEEPIWCVHI2DSMVQWIX3LMV43ASLTON2WKOZSG42TOMJZGY2DKMI>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>