@rschio 2025-01-01
Yes, I did some tests in my branches using my real server and I got really good speedups like ~10x using a concurrent version of ReadAll. That said I think I didn't got the right interface. Probably UnpackBlock shouldn't depend on an internal shared buffer to unpack, I think it should receive a `dst []byte` or a writer...